Have separate version class for jaxrs module
