+++ tessera-jaxrs/common-jaxrs/src/main/java-templates/com/quorum/tessera/api/Version.java
@@ -1,0 +1,8 @@
+package com.quorum.tessera.api;
+
+public class Version {
+
+    public static String getVersion() {
+        return "${project.version}";
+    }
+}
